Refactor day 4
Change core.ai.core to models.core
Change core.ai.history to core.history
